partial bug fix - edge of block terrain type
Mostly fixed a bug that allowed players to avoid the terrain slowdown effects of a block by standing on the edge of the block so that over half the player's body was floating in the air. The fix works most of the time. It doesn't work at X = ±0 and Y = ±0.